# src/deployment/swaggerSchemaAPIGateway.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vianaar",
    "description": "",
    "version": "1.0.0",
    "contact": {
      "name": "",
      "url": "",
      "email": ""
    }
  },
  "components": {
    "securitySchemes": {},
...
  },
  "paths": {
    "/api/v1/projects/{project_id}/units/{unit_id}/book_a_Stay": {
      "post": {
        "tags": ["BookAStay"],
        "description": "API to create book a stay request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "check_in_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "check_out_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "location": {
                    "type": "string"
                  },
                  "no_of_guests": {
                    "type": "number",
                    "minimum": 1
                  }
                },
                "required": ["check_in_date", "check_out_date", "location", "no_of_guests"]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "project_id",
            "required": true
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "unit_id",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "responses": {
            "default": {
              "statusCode": "201"
            }
          },
          "uri": "http://vianaar-backend.eba-hxvhbm2d.ap-south-1.elasticbeanstalk.com/api/v1/projects/{project_id}/units/{unit_id}/book_a_Stay",
          "passthroughBehavior": "when_no_match",
          "httpMethod": "post",
          "type": "http",
          "integrationHttpMethod": "post",
          "requestParameters": {
            "integration.request.path.project_id": "method.request.path.project_id",
            "integration.request.path.unit_id": "method.request.path.unit_id",
            "integration.request.header.Content-Type": "'application/json'"
          }
        }
      },
      "get": {
        "tags": ["BookAStay"],
        "description": "API to fetch current active book a stay request",
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "project_id",
            "required": true
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "unit_id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "check_in_date": {
                      "type": "string",
                      "format": "date"
                    },
                    "check_out_date": {
                      "type": "string",
                      "format": "date"
                    },
                    "location": {
                      "type": "string"
                    },
                    "no_of_guests": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "request_status": {
                      "type": "string"
                    },
                    "add_time": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": ["check_in_date", "check_out_date", "location", "no_of_guests", "add_time"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "responses": {
            "default": {
              "statusCode": "200"
            }
          },
          "uri": "http://vianaar-backend.eba-hxvhbm2d.ap-south-1.elasticbeanstalk.com/api/v1/projects/{project_id}/units/{unit_id}/book_a_Stay",
          "passthroughBehavior": "when_no_match",
          "httpMethod": "GET",
          "type": "http",
          "integrationHttpMethod": "GET",
          "requestParameters": {
            "integration.request.path.project_id": "method.request.path.project_id",
            "integration.request.path.unit_id": "method.request.path.unit_id",
            "integration.request.header.Content-Type": "'application/json'"
          }
        }
      }
    },
...
    "/api/v1/projects/{project_id}/units/{unit_id}/feedback": {
      "post": {
        "tags": ["Feedback"],
        "description": "API to submit feedback",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "comment": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "project_id",
            "required": true
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "unit_id",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "number"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "responses": {
            "default": {
              "statusCode": "201"
            }
          },
          "uri": "http://vianaar-backend.eba-hxvhbm2d.ap-south-1.elasticbeanstalk.com/api/v1/projects/{project_id}/units/{unit_id}/feedback",
          "passthroughBehavior": "when_no_match",
          "httpMethod": "post",
          "type": "http",
          "integrationHttpMethod": "post",
          "requestParameters": {
            "integration.request.path.project_id": "method.request.path.project_id",
            "integration.request.path.unit_id": "method.request.path.unit_id",
            "integration.request.header.Content-Type": "'application/json'"
          }
        }
      },
      "get": {
        "tags": ["Feedback"],
        "description": "API to get feedback",
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "project_id",
            "required": true
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "unit_id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "title": {
                        "type": "string"
                      },
                      "comment": {
                        "type": "string"
                      },
                      "add_time": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "responses": {
            "default": {
              "statusCode": "200"
            }
          },
          "uri": "http://vianaar-backend.eba-hxvhbm2d.ap-south-1.elasticbeanstalk.com/api/v1/projects/{project_id}/units/{unit_id}/feedback",
          "passthroughBehavior": "when_no_match",
          "httpMethod": "get",
          "type": "http",
          "integrationHttpMethod": "get",
          "requestParameters": {
            "integration.request.path.project_id": "method.request.path.project_id",
            "integration.request.path.unit_id": "method.request.path.unit_id",
            "integration.request.header.Content-Type": "'application/json'"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}